move common testutils to pkg, instead of duplicating
