+++ internal/httpserve/handlers/tools_test.go
@@ -2,16 +2,11 @@ package handlers_test
 
 import (
 	"context"
-	"crypto/rand"
-	"crypto/rsa"
-	"io"
 	"log"
 	"path/filepath"
 	"testing"
 	"time"
 
-	"github.com/alicebob/miniredis/v2"
-	"github.com/redis/go-redis/v9"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 
@@ -29,7 +24,6 @@ import (
 	"github.com/datumforge/datum/pkg/middleware/transaction"
 	"github.com/datumforge/datum/pkg/sessions"
 	"github.com/datumforge/datum/pkg/testutils"
-	"github.com/datumforge/datum/pkg/tokens"
 	"github.com/datumforge/datum/pkg/utils/emails"
 	"github.com/datumforge/datum/pkg/utils/marionette"
 )
@@ -152,13 +146,13 @@ func setupEcho(entClient *ent.Client) *echo.Echo {
 
 // handlerSetup to be used for required references in the handler tests
 func handlerSetup(t *testing.T, ent *ent.Client, em *emails.EmailManager, taskMan *marionette.TaskManager) *handlers.Handler {
-	tm, err := createTokenManager(15 * time.Minute) //nolint:mnd
+	tm, err := testutils.CreateTokenManager(15 * time.Minute) //nolint:mnd
 	if err != nil {
 		t.Fatal("error creating token manager")
 	}
 
-	sm := createSessionManager()
-	rc := newRedisClient()
+	sm := testutils.CreateSessionManager()
+	rc := testutils.NewRedisClient()
 	logger := zaptest.NewLogger(t, zaptest.Level(zap.ErrorLevel)).Sugar()
 
 	sessionConfig := sessions.NewSessionConfig(
@@ -189,56 +183,4 @@ func handlerSetup(t *testing.T, ent *ent.Client, em *emails.EmailManager, taskMa
 	}
 
 	return h
-}
-
-func newRedisClient() *redis.Client {
-	mr, err := miniredis.Run()
-	if err != nil {
-		log.Fatalf("an error '%s' was not expected when opening a stub database connection", err)
-	}
-
-	client := redis.NewClient(&redis.Options{
-		Addr:             mr.Addr(),
-		DisableIndentity: true, // # spellcheck:off
-	})
-
-	return client
-}
-
-func createSessionManager() sessions.Store[map[string]any] {
-	hashKey := randomString(32)  //nolint:mnd
-	blockKey := randomString(32) //nolint:mnd
-
-	sm := sessions.NewCookieStore[map[string]any](sessions.DebugCookieConfig,
-		hashKey, blockKey,
-	)
-
-	return sm
-}
-
-func randomString(n int) []byte {
-	id := make([]byte, n)
-
-	if _, err := io.ReadFull(rand.Reader, id); err != nil {
-		panic(err) // This shouldn't happen
-	}
-
-	return id
-}
-
-func createTokenManager(refreshOverlap time.Duration) (*tokens.TokenManager, error) {
-	conf := tokens.Config{
-		Audience:        "http://localhost:17608",
-		Issuer:          "http://localhost:17608",
-		AccessDuration:  1 * time.Hour, //nolint:mnd
-		RefreshDuration: 2 * time.Hour, //nolint:mnd
-		RefreshOverlap:  refreshOverlap,
-	}
-
-	key, err := rsa.GenerateKey(rand.Reader, 2048) //nolint:mnd
-	if err != nil {
-		return nil, err
-	}
-
-	return tokens.NewWithKey(key, conf)
 }
